31770 Object store gallery view not showing thumbnail - Refactored logic to get thumbnail in thumbnail-cell into it's own hook - Reused new hook for StoreObjectGallery to get thumbnail correct - Single point of code control
